Clening up unused code
