+++ bin_types/function.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 
 from urllib3 import Retry
 
-from funcs.blocks_parser import *
+from funcs.brackets_parser import *
 from funcs.exceptions import *
 
 class Function:
