+++ tests_results/expr.txt
@@ -4,4 +4,4 @@ i : [1 1]
 i : [0 1]
 condition : 1
 
-Finished in 0.0029904842376708984 sec
+Finished in 0.0020029544830322266 sec
