+++ tests_results/for.txt
@@ -10,4 +10,4 @@ cont : 0
 cont : 0
 cont : 0
 
-Finished in 0.008976459503173828 sec
+Finished in 0.009383678436279297 sec
