+++ tests_results/func.txt
@@ -5,4 +5,4 @@ side : 1
 else_test_func : 1
 nut : 1
 
-Finished in 0.007978677749633789 sec
+Finished in 0.005998373031616211 sec
