+++ tests_results/operations.txt
@@ -24,4 +24,4 @@ false : 0
 false : 0
 true : 1
 
-Finished in 0.006980180740356445 sec
+Finished in 0.003931522369384766 sec
