+++ tests_results/set_debug.txt
@@ -1,4 +1,4 @@
 
 {'a': '1', 'b': '0', '_list': '[1 0]'}
 
-Finished in 0.0019931793212890625 sec
+Finished in 0.0059816837310791016 sec
