+++ tests_results/std.txt
@@ -16,4 +16,4 @@ eq-0 : 0
 6+11+3= : 20
 5+13+9+2= : 29
 
-Finished in 0.017946481704711914 sec
+Finished in 0.009972572326660156 sec
