+++ tests_results/while.txt
@@ -12,4 +12,4 @@ j : 1
 j : 3
 j : 5
 
-Finished in 0.005983591079711914 sec
+Finished in 0.0069828033447265625 sec
